fix: change float type to double

+++ src/Calculadora.java
@@ -6,16 +6,16 @@
  * @author Caio André
  */
 public class Calculadora {
-  private float a, b, c;
-  private float num1, num2;
+  private double a, b, c;
+  private double num1, num2;
   
   /**
    * Primeiro construtor da classe Calculadora que recebe os parâmetros para o cálculo das raízes reais.
-   * @param a primeiro valor do tipo float
-   * @param b segundo valor do tipo float
-   * @param c terceiro valor do tipo float
+   * @param a primeiro valor do tipo double
+   * @param b segundo valor do tipo double
+   * @param c terceiro valor do tipo double
    */
-  public Calculadora(float a, float b, float c) {
+  public Calculadora(double a, double b, double c) {
     this.a = a;
     this.b = b;
     this.c = c;
@@ -27,7 +27,7 @@ public class Calculadora {
    * @param num2 recebe o segundo valor da operação
    */
   
-  public Calculadora(float num1, float num2){
+  public Calculadora(double num1, double num2){
     this.num1 = num1;
     this.num2 = num2;
   }
@@ -36,7 +36,7 @@ public class Calculadora {
    * Realiza a soma dos valores armazenados nos atributos da classe Calculadora.
    * @return o resultado da operação de soma
    */
-  public float soma(){
+  public double soma(){
     return num1 + num2;
   }
 
@@ -44,7 +44,7 @@ public class Calculadora {
    * Realiza a divisão dos valores armazenados nos atributos da classe Calculadora.
    * @return o resultado da operação de divisão
    */
-  public float divisao(){
+  public double divisao(){
     return num1 / num2;
   }
   
@@ -52,7 +52,7 @@ public class Calculadora {
    * Realiza a multiplicação dos valores armazenados nos atributos da classe Calculadora.
    * @return o resultado da operação de multiplicação
    */
-  public float multiplicacao(){
+  public double multiplicacao(){
     return num1 * num2;
   }
   
@@ -60,7 +60,7 @@ public class Calculadora {
    * Realiza a subtração dos valores armazenados nos atributos da classe Calculadora.
    * @return o resultado da operação de subtração
    */
-  public float subtracao(){
+  public double subtracao(){
     return num1 - num2;
   }
 
@@ -68,7 +68,7 @@ public class Calculadora {
    * Realiza o resto da divisão entre os valores armazenados nos atributos da classe Calculadora.
    * @return o resultado da operação de resto da divisão
    */
-  public float modulo(){
+  public double modulo(){
     return num1 % num2;
   }
 
@@ -77,7 +77,7 @@ public class Calculadora {
    * @return as duas raízes reais resultantes da operação
    */
   public double[] calculaRaizes(){
-    float delta = (b * b) - 4 * a * c;
+    double delta = (b * b) - 4 * a * c;
 
     double x1 = (-b + Math.sqrt(delta)) / (2 * a);
     double x2 = (-b - Math.sqrt(delta)) / (2 * a);
